backend/requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pyyaml
openai
python-dotenv
requests
